# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CIAM Demo Documentation",
...
    "contact": {
      "name": "test",
      "email": "someemail@somedomain.com"
    },
    "version": "v1",
    "x-logo": {
      "url": "https://img.freepik.com/vector-premium/diseno-logotipo-agua-hexagonal-listo-usar_94202-237.jpg?w=360",
      "altText": "Your logo alt text here"
    }
  },
  "paths": {
...
    "/api/Auth/RefreshToken": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Generates a new refresh token.",
        "operationId": "RefreshToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "- The refresh token has al ready expired, back to login (SPCI-401-2).\n- The refresh token is invalid, back to login (SPCI-401-3).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unhandled error, validate the error log.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auth/ChangeUserPassword": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "This endpoint will change the password of the user.",
        "operationId": "ChangeUserPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserChangeResponse"
                }
              }
            }
          },
          "404": {
            "description": "SPCI-404: User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auth/ChangeUserProperty": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "This endpoint will change the property of the user.",
        "operationId": "ChangeUserProperty",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns a 200 if the user property change."
          },
          "404": {
            "description": "SPCI-404: User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auth/UserValidation": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "This endpoint will validate if the user has any of the following states: \n\t\t\t\t\t\t\t\t\t\t\t   <ul><li>NotRegistered = 0</li><li>PreRegistered = 1</li><li>Registered = 2</li>\n\t\t\t\t\t\t\t\t\t\t\t   <li>OTPRequired = 3</li><li>SoftBlock = 4</li><li>HardBlock = 5</li>\n\t\t\t\t\t\t\t\t\t\t\t   <li>CancelledAccount = 8</li><li>FirstLoginOxxo = 20</li></ul>",
        "operationId": "UserValidation",
        "parameters": [
          {
            "name": "PhoneNumber",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": {
              "maxLength": 14,
              "minLength": 10,
              "type": "string",
              "description": "User phone number.",
              "example": "+521234567890"
            }
          },
          {
            "name": "DeviceId",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string",
              "description": "User Device identifier",
              "example": "A1B2C3D4E5"
            }
          },
          {
            "name": "AttemptReference",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "description": "User Attempt Reference",
              "example": "A1B2C3D4E5"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserValidationResponse"
                }
              }
            }
          },
          "404": {
            "description": "SPCI-404: User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auth/GetTransversalUserData": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "Some endpoint description goes here.",
        "operationId": "GetTransversalUserData",
        "parameters": [
          {
            "name": "PhoneNumber",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string",
              "description": "User phone number.",
              "example": "+521234567890"
            }
          },
          {
            "name": "DeviceId",
            "in": "query",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string",
              "description": "Device identifier",
              "example": "A1B2C3D4E5"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDataResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unhandled error, validate the error log.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auth/RevokeRefreshToken": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Some endpoint description goes here.",
        "operationId": "RevokeRefreshToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unhandled error, validate the error log.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auth/GetUserProfile": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "Some endpoint description goes here.",
        "operationId": "GetUserProfile",
        "parameters": [
          {
            "name": "UserId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "description": "User Id",
              "example": "auth0|63b477d3b85540fe7e870500"
            }
          },
          {
            "name": "PhoneNumber",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "description": "User phone number.",
              "example": "+521234567890"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserChangeResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unhandled error, validate the error log.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Auth/UpdateUserProfile": {
      "patch": {
        "tags": [
          "Auth"
        ],
        "description": "Updates the partial information of user.",
        "operationId": "UpdateUserProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PerfilUsuario"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The information is updated correctly",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CambiarCuentaRespuesta"
                }
              }
            }
          },
          "400": {
            "description": "- The information is wrong (SPCI-400): the email provided is not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "- Invalid token (SPCI-401-7): the given token is not valid.\n- Invalid token (SPCI-401-8): the scope header is invalid.\n- Nethone request a review (SPNE-401): review the user.\n- Nethone refuse the request (SPNE-402): user refused.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "- The email has been used. (SPCI-409-2): old email.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "- Loyalty User was not able to create (SPCI-500): if the user not exists in Loyalty.\n- Unhandled error, is necesary validate the log.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auth/RecoveryUserPassword": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Recovery an user's password, based on the requested data.",
        "operationId": "RecoveryUserPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecuperarContrasena"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The password was recovered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CambiarCuentaRespuesta"
                }
              }
            }
          },
          "401": {
            "description": "- Invalid token (SPCI-401-7): the given token is not valid.\n- Invalid token (SPCI-401-8): the scope header is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "- ChangeUserPassword User Not Found in Auth0 (SPCI-500): error. \n- Invalid grant to recover password (SPCI-500): error. \n- RecoveryUserPassword User Not Found in Auth0. (SPCI-500): error.\n- Unhandled error, is necesary validate the log.\n",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auth/CancelUserAccount/CancelUserAccount": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Recovery an user's password, based on the requested data.",
        "operationId": "CancelUserAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelarCuenta"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The account is close correctly.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelarCuentaRespuesta"
                }
              }
            }
          },
          "401": {
            "description": "- The username or password is incorrect (SPCI-401-1): incorrect data (wrong email or password).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "- CancelUserAccount User Not Found in Auth0 (SPCI-500): error.\n- CancelUserAccount invalid grant.(SPCI-500): error.\n- Generic error from Auth0 Service.(SPCI-500): error.\n- Unhandled error, is necesary validate the log.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccessResponse": {
        "required": [
          "accessToken",
          "expiresIn",
          "idToken",
          "refreshToken",
          "tokenType"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "minLength": 1,
            "type": "string",
            "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZ..."
          },
          "tokenType": {
            "minLength": 1,
            "type": "string",
            "example": "Bearer"
          },
          "idToken": {
            "minLength": 1,
            "type": "string",
            "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCIsImtpZ..."
          },
          "expiresIn": {
            "type": "integer",
            "format": "int32",
            "example": 86400
          },
          "refreshToken": {
            "minLength": 1,
            "type": "string",
            "example": "v1.McGUUd_CL7lc4UjVa2jlAgAG7lCyif14ItPqa_VyqXbeLfR1IpHdqu02QT3XuSsGekr0zfh3PJBszwsWGnC8lJw"
          }
        },
        "additionalProperties": false
      },
...
      "CancelarCuenta": {
        "required": [
          "cancellationReason",
          "password",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "minLength": 1,
            "type": "string",
            "description": "User identifier.",
            "example": "auth0|63b477d3b85540fe7e870500"
          },
          "password": {
            "minLength": 1,
            "type": "string",
            "description": "User password.",
            "example": "th1sIsANew%%pwd"
          },
          "cancellationReason": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string",
            "description": "Reason to cancel the account."
          }
        },
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "errorCode": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "httpCode": {
            "$ref": "#/components/schemas/HttpStatusCode"
          },
          "errorCode": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PerfilUsuario": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "User identifier.",
            "nullable": true,
            "example": "auth0|63b477d3b85540fe7e870500"
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Name.",
            "nullable": true,
            "example": "Juan Antonio"
          },
          "surname": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "User surname.",
            "nullable": true,
            "example": "Peréz Pérez"
          },
          "birthday": {
            "type": "string",
            "description": "User birthday.",
            "nullable": true,
            "example": "01/01/0001"
          },
          "email": {
            "type": "string",
            "description": "User email.",
            "nullable": true,
            "example": "examplemail@cool.domain.com"
          },
          "zipCode": {
            "type": "string",
            "description": "User zipCode.",
            "nullable": true,
            "example": "64000"
          },
          "secondLastName": {
            "type": "string",
            "description": "User second last name.",
            "nullable": true,
            "example": "Peréz"
          },
          "gender": {
            "pattern": "[M|W]",
            "type": "string",
            "description": "User gender.",
            "nullable": true,
            "example": "M (for Men) or W (for Woman)"
          },
          "attemptReference": {
            "type": "string",
            "description": "User attempt reference.",
            "nullable": true,
            "example": "A1B2C3D4E5"
          },
          "deviceId": {
            "type": "string",
            "description": "User device identifier.",
            "nullable": true,
            "example": "x1234ABCD"
          }
        },
        "additionalProperties": false,
        "description": ""
      },
      "RecuperarContrasena": {
        "required": [
          "newPassword"
        ],
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "description": "User phone number.",
            "nullable": true,
            "example": "+525544332211"
          },
          "newPassword": {
            "maxLength": 50,
            "minLength": 8,
            "type": "string",
            "description": "User new password to update.",
            "example": "th1sIsANew%%pwd"
          },
          "deviceId": {
            "type": "string",
            "description": "User device identifier.",
            "nullable": true,
            "example": "a3f3a772-f9af-4162-8ac3-8d79ad"
          },
          "attemptReference": {
            "type": "string",
            "description": "User attempt reference.",
            "nullable": true,
            "example": "A1B2C3D4E5"
          }
        },
        "additionalProperties": false
      },
      "RefreshRequest": {
        "required": [
          "refreshToken"
        ],
        "type": "object",
        "properties": {
          "refreshToken": {
            "minLength": 1,
            "type": "string",
            "description": "Refresh access token given.",
            "example": "v1.McGUUd_CL7lc4UjVa2jlAgAG7lCyif14ItPqa_VyqXbeLfR1IpHdqu02QT3XuSsGekr0zfh3PJBszwsWGnC8lJw"
          }
        },
        "additionalProperties": false
      },
      "RefreshResponse": {
        "required": [
          "refreshToken"
        ],
        "type": "object",
        "properties": {
          "refreshToken": {
            "minLength": 1,
            "type": "string",
            "description": "Refresh access token given.",
            "example": "v1.McGUUd_CL7lc4UjVa2jlAgAG7lCyif14ItPqa_VyqXbeLfR1IpHdqu02QT3XuSsGekr0zfh3PJBszwsWGnC8lJw"
          }
        },
        "additionalProperties": false
      },
...
      "UpdatePasswordRequest": {
        "required": [
          "newPassword",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "minLength": 1,
            "type": "string",
            "description": "User Id Auth0",
            "example": "auth0|63b477d3b85540fe7e870500"
          },
          "newPassword": {
            "maxLength": 50,
            "minLength": 8,
            "type": "string",
            "description": "User New password",
            "example": "th1sIsANew%%pwd"
          },
          "currentPassword": {
            "maxLength": 50,
            "minLength": 8,
            "type": "string",
            "description": "User Current password",
            "nullable": true,
            "example": "th1sIsAOld%%pwd"
          }
        },
        "additionalProperties": false
      },
      "UpdateUserRequest": {
        "required": [
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "minLength": 1,
            "type": "string",
            "description": "User Id Auth0",
            "example": "auth0|63b477d3b85540fe7e870500"
          },
          "email": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "User email.",
            "nullable": true,
            "example": "examplemail@cool.domain.com"
          },
          "phoneNumber": {
            "type": "string",
            "description": "User phone number.",
            "nullable": true,
            "example": "+521234567890"
          }
        },
        "additionalProperties": false
      },
      "UserChangeResponse": {
        "required": [
          "birthday",
          "emailVerified",
          "firstName",
          "lastName",
          "loyaltyId",
          "phoneNumber",
          "zipCode"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email.",
            "nullable": true,
            "example": "+521234567890"
          },
          "userName": {
            "type": "string",
            "description": "User Name of user",
            "nullable": true,
            "example": "+521234567890"
          },
          "firstName": {
            "minLength": 1,
            "type": "string",
            "description": "User first Name",
            "example": "Juan Antonio"
          },
          "lastName": {
            "minLength": 1,
            "type": "string",
            "description": "User last Name.",
            "example": "Peréz Pérez"
          },
          "birthday": {
            "minLength": 1,
            "type": "string",
            "description": "User birthday.",
            "example": "01/01/2000"
          },
          "phoneNumber": {
            "minLength": 1,
            "type": "string",
            "description": "User phone number.",
            "example": "+525555555555"
          },
          "zipCode": {
            "minLength": 1,
            "type": "string",
            "description": "User zip code",
            "example": "64000"
          },
          "loyaltyId": {
            "minLength": 1,
            "type": "string",
            "description": "Loyalty Id of the user",
            "example": "P1ATYGR"
          },
          "emailVerified": {
            "type": "boolean",
            "description": "Check if the user's email is verified",
            "example": "False"
          }
        },
        "additionalProperties": false
      },
      "UserDataResponse": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "User first Name",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "description": "User last Name.",
            "nullable": true
          },
          "email": {
            "type": "string",
            "description": "User email.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserValidationResponse": {
        "required": [
          "codeUserStatus"
        ],
        "type": "object",
        "properties": {
          "codeUserStatus": {
            "minLength": 1,
            "type": "string",
            "description": "User Status",
            "example": "0"
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}